LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerrawupdate_remove_target.https-expected.txt



FAIL "pointermove" and its preceding boundary events should be fired on parent if "pointerrawupdate" event listener removes its target assert_equals: expected "pointerover@target,pointerrawupdate@target,pointerover@container,pointermove@container" but got "pointerover@target,pointermove@target"
FAIL "pointermove" and its preceding boundary events should be fired on ancestor if "pointerrawupdate" event listener removes its target parent assert_equals: expected "pointerover@target,pointerrawupdate@target,pointerover@container,pointermove@container" but got "pointerover@target,pointermove@target"
FAIL "pointermove" and its preceding boundary events should be fired on parent if "pointerrawupdate" event listener removes its document assert_equals: expected "pointerover@target, pointerrawupdate@target, pointerover@container, pointermove@container" but got "pointerover@container, pointermove@container"
